# internal/server/api/openapi.gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.7.0 DO NOT EDIT.
package api

import (
	"bytes"
	"context"
	"encoding/json"
	"errors"
	"fmt"
	"io"
	"net/http"

	"github.com/go-chi/chi/v5"
	"github.com/oapi-codegen/runtime"
	openapi_types "github.com/oapi-codegen/runtime/types"
)

// Error defines model for Error.
type Error struct {
	Error string `json:"error"`
}

// FetchFailure defines model for FetchFailure.
type FetchFailure struct {
	Currency string `json:"currency"`
	Error    string `json:"error"`
}

// FetchRatesRequest defines model for FetchRatesRequest.
type FetchRatesRequest struct {
	Currencies *[]string `json:"currencies,omitempty"`
}

// FetchRatesResponse defines model for FetchRatesResponse.
type FetchRatesResponse struct {
	Failed []FetchFailure `json:"failed"`
	Saved  []Rate         `json:"saved"`
}

// Rate defines model for Rate.
type Rate struct {
	Currency      string             `json:"currency"`
	EffectiveDate openapi_types.Date `json:"effective_date"`
	Id            *string            `json:"id,omitempty"`
	Rate          string             `json:"rate"`
	Source        string             `json:"source"`
}

// RatesResponse defines model for RatesResponse.
type RatesResponse struct {
	BaseCurrency string `json:"base_currency"`
	Rates        []Rate `json:"rates"`
}

// GetRatesCurrencyHistoryParams defines parameters for GetRatesCurrencyHistory.
type GetRatesCurrencyHistoryParams struct {
	From  *openapi_types.Date `form:"from,omitempty" json:"from,omitempty"`
	To    *openapi_types.Date `form:"to,omitempty" json:"to,omitempty"`
	Limit *int                `form:"limit,omitempty" json:"limit,omitempty"`
}

// PostRatesFetchJSONRequestBody defines body for PostRatesFetch for application/json ContentType.
type PostRatesFetchJSONRequestBody = FetchRatesRequest

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Latest exchange rates
	// (GET /rates)
	GetRates(w http.ResponseWriter, r *http.Request)
	// Fetch current exchange rates
	// (POST /rates/fetch)
	PostRatesFetch(w http.ResponseWriter, r *http.Request)
	// Exchange-rate history for one currency
	// (GET /rates/{currency}/history)
	GetRatesCurrencyHistory(w http.ResponseWriter, r *http.Request, currency string, params GetRatesCurrencyHistoryParams)
}

// Unimplemented server implementation that returns http.StatusNotImplemented for each endpoint.

type Unimplemented struct{}

// Latest exchange rates
// (GET /rates)
func (_ Unimplemented) GetRates(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Fetch current exchange rates
// (POST /rates/fetch)
func (_ Unimplemented) PostRatesFetch(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Exchange-rate history for one currency
// (GET /rates/{currency}/history)
func (_ Unimplemented) GetRatesCurrencyHistory(w http.ResponseWriter, r *http.Request, currency string, params GetRatesCurrencyHistoryParams) {
	w.WriteHeader(http.StatusNotImplemented)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// GetRates operation middleware
func (siw *ServerInterfaceWrapper) GetRates(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetRates(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// PostRatesFetch operation middleware
func (siw *ServerInterfaceWrapper) PostRatesFetch(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.PostRatesFetch(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetRatesCurrencyHistory operation middleware
func (siw *ServerInterfaceWrapper) GetRatesCurrencyHistory(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "currency" -------------
	var currency string

	err = runtime.BindStyledParameterWithOptions("simple", "currency", chi.URLParam(r, "currency"), &currency, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "currency", Err: err})
		return
	}

	// Parameter object where we will unmarshal all parameters from the context
	var params GetRatesCurrencyHistoryParams

	// ------------- Optional query parameter "from" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "from", r.URL.Query(), &params.From, runtime.BindQueryParameterOptions{Type: "string", Format: "date"})
	if err != nil {
		var requiredError *runtime.RequiredParameterError
		if errors.As(err, &requiredError) {
			siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "from"})
		} else {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "from", Err: err})
		}
		return
	}

	// ------------- Optional query parameter "to" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "to", r.URL.Query(), &params.To, runtime.BindQueryParameterOptions{Type: "string", Format: "date"})
	if err != nil {
		var requiredError *runtime.RequiredParameterError
		if errors.As(err, &requiredError) {
			siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "to"})
		} else {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "to", Err: err})
		}
		return
	}

	// ------------- Optional query parameter "limit" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "limit", r.URL.Query(), &params.Limit, runtime.BindQueryParameterOptions{Type: "integer", Format: ""})
	if err != nil {
		var requiredError *runtime.RequiredParameterError
		if errors.As(err, &requiredError) {
			siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "limit"})
		} else {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "limit", Err: err})
		}
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetRatesCurrencyHistory(w, r, currency, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{})
}

type ChiServerOptions struct {
	BaseURL          string
	BaseRouter       chi.Router
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, r chi.Router) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseRouter: r,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, r chi.Router, baseURL string) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseURL:    baseURL,
		BaseRouter: r,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options ChiServerOptions) http.Handler {
	r := options.BaseRouter

	if r == nil {
		r = chi.NewRouter()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}
	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/rates", wrapper.GetRates)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/rates/fetch", wrapper.PostRatesFetch)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/rates/{currency}/history", wrapper.GetRatesCurrencyHistory)
	})

	return r
}

type GetRatesRequestObject struct {
}

type GetRatesResponseObject interface {
	VisitGetRatesResponse(w http.ResponseWriter) error
}

type GetRates200JSONResponse RatesResponse

func (response GetRates200JSONResponse) VisitGetRatesResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)
	_, err := buf.WriteTo(w)
	return err
}

type PostRatesFetchRequestObject struct {
	Body *PostRatesFetchJSONRequestBody
}

type PostRatesFetchResponseObject interface {
	VisitPostRatesFetchResponse(w http.ResponseWriter) error
}

type PostRatesFetch200JSONResponse FetchRatesResponse

func (response PostRatesFetch200JSONResponse) VisitPostRatesFetchResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)
	_, err := buf.WriteTo(w)
	return err
}

type PostRatesFetch207JSONResponse FetchRatesResponse

func (response PostRatesFetch207JSONResponse) VisitPostRatesFetchResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(207)
	_, err := buf.WriteTo(w)
	return err
}

type PostRatesFetch500JSONResponse Error

func (response PostRatesFetch500JSONResponse) VisitPostRatesFetchResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)
	_, err := buf.WriteTo(w)
	return err
}

type GetRatesCurrencyHistoryRequestObject struct {
	Currency string `json:"currency"`
	Params   GetRatesCurrencyHistoryParams
}

type GetRatesCurrencyHistoryResponseObject interface {
	VisitGetRatesCurrencyHistoryResponse(w http.ResponseWriter) error
}

type GetRatesCurrencyHistory200JSONResponse RatesResponse

func (response GetRatesCurrencyHistory200JSONResponse) VisitGetRatesCurrencyHistoryResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)
	_, err := buf.WriteTo(w)
	return err
}

type GetRatesCurrencyHistory422JSONResponse Error

func (response GetRatesCurrencyHistory422JSONResponse) VisitGetRatesCurrencyHistoryResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(422)
	_, err := buf.WriteTo(w)
	return err
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {
	// Latest exchange rates
	// (GET /rates)
	GetRates(ctx context.Context, request GetRatesRequestObject) (GetRatesResponseObject, error)
	// Fetch current exchange rates
	// (POST /rates/fetch)
	PostRatesFetch(ctx context.Context, request PostRatesFetchRequestObject) (PostRatesFetchResponseObject, error)
	// Exchange-rate history for one currency
	// (GET /rates/{currency}/history)
	GetRatesCurrencyHistory(ctx context.Context, request GetRatesCurrencyHistoryRequestObject) (GetRatesCurrencyHistoryResponseObject, error)
}

type StrictHandlerFunc func(ctx context.Context, w http.ResponseWriter, r *http.Request, request any) (any, error)
type StrictMiddlewareFunc func(f StrictHandlerFunc, operationID string) StrictHandlerFunc

type StrictHTTPServerOptions struct {
	RequestErrorHandlerFunc  func(w http.ResponseWriter, r *http.Request, err error)
	ResponseErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: StrictHTTPServerOptions{
		RequestErrorHandlerFunc: func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		},
		ResponseErrorHandlerFunc: func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusInternalServerError)
		},
	}}
}

func NewStrictHandlerWithOptions(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc, options StrictHTTPServerOptions) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: options}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
	options     StrictHTTPServerOptions
}

// GetRates operation middleware
func (sh *strictHandler) GetRates(w http.ResponseWriter, r *http.Request) {
	var request GetRatesRequestObject

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.GetRates(ctx, request.(GetRatesRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetRates")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(GetRatesResponseObject); ok {
		if err := validResponse.VisitGetRatesResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// PostRatesFetch operation middleware
func (sh *strictHandler) PostRatesFetch(w http.ResponseWriter, r *http.Request) {
	var request PostRatesFetchRequestObject

	var body PostRatesFetchJSONRequestBody
	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
		if !errors.Is(err, io.EOF) {
			sh.options.RequestErrorHandlerFunc(w, r, fmt.Errorf("can't decode JSON body: %w", err))
			return
		}
	} else {
		request.Body = &body
	}

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.PostRatesFetch(ctx, request.(PostRatesFetchRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "PostRatesFetch")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(PostRatesFetchResponseObject); ok {
		if err := validResponse.VisitPostRatesFetchResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// GetRatesCurrencyHistory operation middleware
func (sh *strictHandler) GetRatesCurrencyHistory(w http.ResponseWriter, r *http.Request, currency string, params GetRatesCurrencyHistoryParams) {
	var request GetRatesCurrencyHistoryRequestObject

	request.Currency = currency
	request.Params = params

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.GetRatesCurrencyHistory(ctx, request.(GetRatesCurrencyHistoryRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetRatesCurrencyHistory")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(GetRatesCurrencyHistoryResponseObject); ok {
		if err := validResponse.VisitGetRatesCurrencyHistoryResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}